+++ app/controllers/carto/api/users_controller.rb
@@ -1,13 +1,40 @@
 module Carto
   module Api
     class UsersController < ::Api::ApplicationController
+      include AppAssetsHelper
+      include MapsApiHelper
+      include SqlApiHelper
+      include CartoDB::ConfigUtils
+      include FrontendConfigHelper
 
-      ssl_required :get_authenticated_users, :show
+      ssl_required :show, :me, :get_authenticated_users
 
       skip_before_filter :api_authorization_required, only: [:get_authenticated_users]
 
       def show
         render json: Carto::Api::UserPresenter.new(uri_user).data
+      end
+
+      def me
+        carto_viewer = Carto::User.find(current_viewer.id)
+        dashboard_notifications = carto_viewer.notifications_for_category(:dashboard)
+        organization_notifications = carto_viewer.received_notifications.unread.map do |n|
+          Carto::Api::ReceivedNotificationPresenter.new(n)
+        end
+
+        render json: {
+          username: current_user.username,
+          user_data: current_user.data,
+          config: frontend_config_hash,
+          upgrade_url: current_user.upgrade_url(request.protocol),
+          isFirstTimeViewingDashboard: !current_user.dashboard_viewed?,
+          isJustLoggedIn: !!flash['logged'],
+          default_fallback_basemap: current_user.default_basemap,
+          dashboard_notifications: dashboard_notifications,
+          organization_notifications: organization_notifications,
+          user_frontend_version: current_user.user_frontend_version,
+          asset_host: current_user.asset_host
+        }
       end
 
       def get_authenticated_users