+++ NEWS.md
@@ -2,6 +2,9 @@ Development
 -----------
 
 ### Features
+* Split frontend for Dashboard view from Rails backend (#12607)
+* Add /api/v3/me endpoint (#12599)
+
 * Sidebar UI tweaks (#12479)
 * Improving affordance of Delete icon (#12531)
 * Change select "attribute" placeholders (#12498)